+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Login page UX improvements
+- Minor refactor of praise/controllers/quantify to clarify logic flow. #221
 
 ## [0.2.0] - 2022-03-31